Fix failing specs

+++ db/seeds.rb
@@ -3,8 +3,11 @@ def seed_from_file(filename)
   seeds.map { |seed| yield(seed.chomp) if block_given? }
 end
 
-BlacklistedPassword.delete_all
-DefaultGroupCover.delete_all
 
+BlacklistedPassword.delete_all
 seed_from_file('password_blacklist') { |password| BlacklistedPassword.create(string: password) }
-seed_from_file('default_group_covers') { |url| DefaultGroupCover.store(url) }
+
+if !Rails.env.test?
+  DefaultGroupCover.delete_all
+  seed_from_file('default_group_covers') { |url| DefaultGroupCover.store(url) }
+end
